refactor: add winning lotto ticket to WinningLotto file
